pyproject.toml
[project]
name = "framework-benchmark"
version = "0.1.0"
description = "Benchmark comparing Django-Bolt, Django Ninja, FastAPI, and Litestar"
requires-python = ">=3.12"
dependencies = [
    # FastAPI
    "fastapi>=0.127.0",
    "uvicorn[standard]>=0.32.0",
    # Litestar
    "litestar[standard]>=2.19.0",
    # Django + Django Ninja + Django Bolt + Django REST Framework
    "django>=5.0",
    "django-ninja>=1.5.1",
    "django-bolt==0.4.7",
    "djangorestframework>=3.15.0",
    # GraphQL (ASGI)
    "graphene-django>=3.2.0",
    "strawberry-graphql[fastapi]>=0.237.0",
    "strawberry-graphql-django>=0.23.0",
    "starlette>=0.36.0",
    # Async SQLite for FastAPI ORM tests
    "sqlalchemy[asyncio]>=2.0.0",
    "aiosqlite>=0.20.0",
    # Shared
    "msgspec>=0.19.0",
    "orjson>=3.10.0",
    # Graphing
    "matplotlib>=3.9.0",
]

[project.optional-dependencies]
dev = [
    "httpx>=0.28.0",
    "pytest>=8.0.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["shared", "django_project"]
